requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Flask==2.3.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.0
requests==2.31.0
urllib3==2.4.0
Werkzeug==3.1.3
taxii2-client==2.3.0
stix2==3.0.1
pandas==2.2.2


